ci: first-run check

+++ web/build.rs
@@ -7,9 +7,11 @@ fn main() {
     fs::copy("../browser-ext/script.js", "../static/script.js")
         .expect("copy script file from `browser-ext` folder to static folder");
 
-    fs::copy(
-        "../browser-ext/script.js",
-        "../frontend/public/app/script.js",
-    )
-    .expect("copy script file from `browser-ext` folder to frontend folder");
+    if Path::new("../browser-ext/").exists() {
+        fs::copy(
+            "../browser-ext/script.js",
+            "../frontend/public/app/script.js",
+        )
+        .expect("copy script file from `browser-ext` folder to frontend folder");
+    }
 }
